adding SP name to teraform create

+++ Create_Secure_Perimeter/create_secure_perimeter.sh
@@ -9,25 +9,35 @@
 #
 ################################################################
 
+if [ $# -ne 1 ]
+  then
+    echo "Usage : create_secure_perimeter.sh <Unique name of terraform workspace - for example SP1>"
 
-cd secure_perimeter
-# Remove these files as they may have been copied in during previous eployment - 
-# They cannot be deployed in first stage of deploying secure perimeter
-rm -f ./modules/monitoring_cluster/network_pod.tf
-rm -f ./modules/monitoring_cluster/health_pod.tf
+else
+   workspace_name=$1
 
-terraform init
-terraform apply
+   cd secure_perimeter
+   terraform workspace new ${workspace_name}
+   # Remove these files as they may have been copied in during previous deployment - 
+   # They cannot be deployed in first stage of deploying secure perimeter
+
+   rm -f ./modules/monitoring_cluster/network_pod.tf
+   rm -f ./modules/monitoring_cluster/health_pod.tf
+
+   terraform init
+   terraform apply
 
 
-#wait for monitoring cluster to come up
+   #wait for monitoring cluster to come up
 
-cp ../health_pod.tf ./modules/monitoring_cluster
-terraform init
-terraform apply  -target=module.monitoring_cluster.module.health-pod -auto-approve
+   cp ../health_pod.tf ./modules/monitoring_cluster
+   terraform init
+   terraform apply  -target=module.monitoring_cluster.module.health-pod -auto-approve
 
-cp ../network_pod.tf ./modules/monitoring_cluster
-terraform init
-terraform apply  -target=module.monitoring_cluster.module.network-pod -auto-approve
+   cp ../network_pod.tf ./modules/monitoring_cluster
+   terraform init
+   terraform apply  -target=module.monitoring_cluster.module.network-pod -auto-approve
 
+   
+fi
